Guardrails-Implementation-in-LLMs/requirements.txt
alt-profanity-check
openai
guardrails-ai
python-dotenv
streamlit  